Add theme styles and update components
Introduced a theme.styled.tsx file for defining global styles across different components and made changes to button.styled.tsx for centralized styling. The App.tsx file was updated to make use of the new theme styles. This approach ensures a consistent look and feel across the application and simplifies the maintenance of the code by centralizing the styles in a single file. This also resulted in the removal of some individual style definitions from the various components.